renamed builtins to core_plugins.
This is part of a refactoring to move out more core commands to plugins.
