+++ errbot/core_plugins/__init__.py
