+++ errbot/core_plugins/chatRoom.plug
