+++ errbot/core_plugins/chatRoom.py
