+++ errbot/core_plugins/templates/__init__.py
