+++ errbot/core_plugins/vcheck.plug
