+++ errbot/core_plugins/vcheck.py
