+++ errbot/core_plugins/wsview.py
